+++ MRP_Core.lua
@@ -960,6 +960,8 @@ function Core:HandleSlashCommand(msg)
         if MRP.Changelog then
             MRP.Changelog:Show()
         end
+    elseif cmd == "update" or cmd == "updates" then
+        MRP.Update:Show()
     elseif cmd == "trashit" then -- MRP_REMOVE_LINE
         Core:TrashIt()           -- MRP_REMOVE_LINE
     elseif cmd == "route" then
@@ -975,6 +977,7 @@ function Core:HandleSlashCommand(msg)
         print(L[" - /mrp waypoint none|waypoint|tomtom → Choose the waypoint system"])
         print(L[" - /mrp tomtom on|off → (Deprecated) Use /mrp waypoint instead"])
         print(L[" - /mrp changelog → Open the latest changelog popup"])
+        print(L[" - /mrp update → Show official download sources"])
         print(L[" - /mrp route → Recalculate the optimized route"])
         print(L[" - /mrp updatedisplaydelayed <number> → Force update of the display after a delay (default: 0.25 seconds)"])
     end